[#15] Backend API code clean up -Changed protected methods to private methods -moved factService to services directory -moved staticStrings to services directory -added file upload folder to .gitignore
