+++ src/backend_service/.gitignore
@@ -5,3 +5,6 @@ venv
 .cache
 .coverage
 .idea/*
+
+# File upload directory
+uploads/
