+++ src/backend_service/services/staticStrings.py
